🧼 rename notes to actions
this just makes it much more clear what should go in here. this
directory is where the actions go and all of the functionality is
stored in helpers/utils.
