+++ src/actions/index.ts
@@ -1,5 +1,4 @@
 export * from "./interactiveNotePrompt"
 export * from "./removeNote"
 export * from "./newNote"
-export * from "./noteHelpers"
 export * from "./updateNote"
