+++ src/utils/index.ts
@@ -2,4 +2,5 @@ export * from "./config"
 export * from "./dateHelpers"
 export * from "./contentHelpers"
 export * from "./notes"
+export * from "./noteHelpers"
 export * from "./noteDates"
